Additional GUI Features

+++ app/src/main/java/group9/project/PaneView.java
@@ -4,7 +4,7 @@ import javafx.scene.layout.Pane;
 
 public abstract class PaneView extends Pane implements IStartable, IUpdateable
 {
-    protected int width, height;
+    protected double width, height;
 
     public PaneView(int newWidth, int newHeight)
     {
